HFP-120 [feat] 마이페이지 프리랜서 (고용주 리뷰) 점수 API 구현)

+++ freebridge/user/src/main/java/com/fallguys/mypage/entity/freelancer/Collaboration.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Collaboration {
-    private Integer communication;
-    private Integer scheduleAdherence;
-    private Integer dispute;
+    private Double communication;
+    private Double scheduleAdherence;
+    private Double dispute;
 }
